added all screen

+++ app/src/main/java/com/example/testapplication/cloud/model/ImagesAPI.kt
@@ -6,12 +6,12 @@ import com.google.gson.annotations.SerializedName
 class ImagesAPI {
     @SerializedName("id")
     @Expose
-    var id: Int? = null
+    var id: Long? = null
         private set
 
     @SerializedName("product_id")
     @Expose
-    var productId: Int? = null
+    var productId: Long? = null
         private set
 
     @SerializedName("position")
